api/requirements.txt
azure-functions==1.16.0
pandas==2.2.3
mysql-connector-python==8.1.0
python-dotenv==1.0.0
openpyxl==3.1.2